+++ example/lib/common/utils/crop_editor_helper.dart
@@ -194,7 +194,7 @@ Future<List<int>> cropImageDataWithNativeLibrary(
 /// it may be failed, due to Cross-domain
 Future<Uint8List> _loadNetwork(ExtendedNetworkImageProvider key) async {
   try {
-    final Response response = await HttpClientHelper.get(key.url,
+    final Response response = await HttpClientHelper.get(Uri.parse(key.url),
         headers: key.headers,
         timeLimit: key.timeLimit,
         timeRetry: key.timeRetry,